+++ backend/tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/index.ts","./src/products.ts","./src/database/db.ts","./src/database/dbtypes.ts","./src/middleware/adminmiddleware.ts","./src/middleware/authmiddleware.ts","./src/middleware/tokenverify.ts","./src/routes/adminroutes.ts","./src/routes/products.ts","./src/routes/userauth.ts","./src/validations/products.ts","./src/validations/uservalidation.ts"],"version":"5.7.2"}
+{"root":["./src/app.ts","./src/index.ts","./src/products.ts","./src/controllers/admincontroller.ts","./src/controllers/productcontroller.ts","./src/controllers/usercontroller.ts","./src/database/dbconnection.ts","./src/database/dbtypes.ts","./src/middleware/adminmiddleware.ts","./src/middleware/authmiddleware.ts","./src/middleware/tokenverify.ts","./src/models/admin.ts","./src/models/cart.ts","./src/models/order.ts","./src/models/product.ts","./src/models/user.ts","./src/models/index.ts","./src/routes/adminroutes.ts","./src/routes/products.ts","./src/routes/userauth.ts","./src/validations/products.ts","./src/validations/uservalidation.ts"],"version":"5.7.2"}
